+++ fiddles/nativeScript/fiddle-0010-Groceries/groceries/app/app.component.js
@@ -1,0 +1,20 @@
+"use strict";
+var core_1 = require("@angular/core");
+var AppComponent = (function () {
+    function AppComponent() {
+    }
+    AppComponent.prototype.submit = function () {
+        console.log("hello");
+    };
+    AppComponent = __decorate([
+        core_1.Component({
+            selector: "my-app",
+            templateUrl: "app.component.html",
+            styleUrls: ["pages/login/login-common.css", "pages/login/login.css"]
+        }), 
+        __metadata('design:paramtypes', [])
+    ], AppComponent);
+    return AppComponent;
+}());
+exports.AppComponent = AppComponent;
+//# sourceMappingURL=app.component.js.map