+++ fiddles/nativeScript/fiddle-0010-Groceries/groceries/app/app.component.ts
@@ -1,0 +1,12 @@
+import {Component} from "@angular/core";
+
+@Component({
+    selector: "my-app",
+    templateUrl: "app.component.html",
+    styleUrls: ["pages/login/login-common.css", "pages/login/login.css"]
+})
+export class AppComponent {
+    submit() {
+        console.log("hello");
+    }
+}