+++ fiddles/nativeScript/fiddle-0010-Puzzle/puzzle/tools/env/env-config.interface.ts
@@ -1,2 +1,0 @@
-export { EnvConfig } from '../../src/client/app/frameworks/core/utils/config';
-